refactor: delete useless setUserData helper
