+++ src/modules/UserProfile/types.ts
@@ -6,7 +6,6 @@ export interface IFieldInfo {
   validation:
     | ObjectSchema<Record<string, string>>
     | ObjectSchema<Record<string, Date>>;
-  // setUserData: Dispatch<SetStateAction<Record<string, string | undefined>>>;
 }
 
-export type IFieldData = Omit<IFieldInfo, 'setUserData'>;
+export type IFieldData = IFieldInfo;
