--- src/store/userDataSlice/helpers/index.ts
@@ -1,4 +1,0 @@
-import setUserData from './setUserData';
-import getRegisteredUserData from './getRegisteredUserData';
-
-export { setUserData, getRegisteredUserData };
